requirements.txt
streamlit==1.45.1
pyyaml==6.0.2
streamlit-extras==0.7.1
shap==0.47.2
joblib==1.5.1
numpy==2.2.6
llama-cpp-python==0.3.9
pandas==2.2.3
scikit-learn==1.6.1
huggingface_hub==0.32.3
google-genai==1.18.0
python-dotenv==1.1.0
